Implemented TCP networking clients and system-clock timers
The meter and utility network clients can now be implemented with
TcpNetworkClient and TcpUtilityClient, which use epoll() to multiplex
several incoming connections on a single socket without adding threads.
The TimerManager feature can now be imlemented with LinuxTimerManager,
which uses the system clock to set timeouts, which is necessary in order
to use the TCP networking clients (since they need to use real time to
measure timeouts, not event simulator clock ticks).

Also, I changed the build system to use an explicit Makefile instead of
relying on Eclipse's CDT builder to figure it out, because now there are
multiple possible "main" files depending on whether we wish to compile
in simulated mode or in "real networking" mode.

+++ src/CrusaderAgreementState.cpp
@@ -32,7 +32,7 @@ namespace pddm {
 std::vector<std::shared_ptr<messaging::OverlayMessage> > CrusaderAgreementState::finish_phase_1(int current_round) {
     std::vector<std::shared_ptr<messaging::OverlayMessage>> accept_messages;
     for(const auto& signed_value_entry : signed_proxy_values) {
-        if(signed_value_entry.second.signatures.size() < log2n + 1) {
+        if(signed_value_entry.second.signatures.size() < (unsigned) log2n + 1) {
             //Reject values without enough signatures
             continue;
         }
@@ -61,7 +61,7 @@ util::unordered_ptr_set<messaging::ValueContribution> CrusaderAgreementState::fi
     util::unordered_ptr_set<messaging::ValueContribution> accepted_proxy_values;
     for(const auto& signed_value_entry : signed_proxy_values) {
         //Accept the value if it has enough distinct signatures
-        if(signed_value_entry.second.signatures.size() >= log2n + 1) {
+        if(signed_value_entry.second.signatures.size() >= (unsigned) log2n + 1) {
             accepted_proxy_values.emplace(signed_value_entry.second.value);
         } else {
             //Log warning
